+++ .gitignore
@@ -1,3 +1,7 @@
+# Project Based
+*.db
+config.json
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
