+++ comet/templates/index.html
@@ -481,7 +481,7 @@
             </div>
 
             <div class="form-item">
-                <sl-input id="maxResults" type="number" min=0 value=0 label="Max Results" placeholder="Enter max results"></sl-input>
+                <sl-input id="maxResults" type="number" min=0 value=10 label="Max Results" placeholder="Enter max results"></sl-input>
             </div>
 
             <div class="form-item">
