+++ comet/utils/logger.py
@@ -1,7 +1,12 @@
 import sys
+
 from loguru import logger
 
+
 def setupLogger(level: str):
+    logger.level("COMET", no=50, icon="🌠", color="<fg #7871d6>")
+    logger.level("API", no=40, icon="👾", color="<fg #7871d6>")
+
     logger.level("INFO", icon="📰", color="<fg #FC5F39>")
     logger.level("DEBUG", icon="🕸️", color="<fg #DC5F00>")
     logger.level("WARNING", icon="⚠️", color="<fg #DC5F00>")
