+++ pyproject.toml
@@ -19,7 +19,12 @@ asyncio = "^3.4.3"
 loguru = "^0.7.2"
 databases = "^0.9.0"
 aiosqlite = "^0.20.0"
+pydantic-settings = "^2.3.4"
 
+[tool.poetry.group.dev.dependencies]
+isort = "^5.13.2"
+pyright = "^1.1.369"
+pytest = "^8.2.2"
 
 [build-system]
 requires = ["poetry-core"]
